backend/requirements/base.txt
alabaster==1.0.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
asttokens==3.0.1
asyncpg==0.31.0
attrs==25.4.0
autobahn==25.12.2
automat==25.4.16
autopep8==2.3.2
babel==2.18.0
bcrypt==5.0.0
billiard==4.2.4
black==26.1.0
cbor2==5.8.0
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
channels==4.3.2
channels-redis==4.3.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.5
daphne==4.2.1
databases==0.9.0
decorator==5.2.1
django==4.2.29
django-cors-headers==4.9.0
djangorestframework==3.16.1
docutils==0.22.4
ecdsa==0.19.1
executing==2.2.1
factory-boy==3.3.3
faker==40.5.1
fastapi==0.135.1
flake8==7.3.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
imagesize==2.0.0
incremental==24.11.0
iniconfig==2.3.0
ipdb==0.13.13
ipython==9.10.0
ipython-pygments-lexers==1.1.1
isort==8.0.1
jedi==0.19.2
jinja2==3.1.6
kombu==5.6.2
librt==0.8.1
markupsafe==3.0.3
matplotlib-inline==0.2.1
mccabe==0.7.0
msgpack==1.1.2
mypy==1.19.1
mypy-extensions==1.1.0
packaging==26.0
parso==0.8.6
passlib==1.7.4
pathspec==1.0.4
platformdirs==4.9.2
pluggy==1.6.0
prompt-toolkit==3.0.52
psycopg2-binary==2.9.11
pure-eval==0.2.3
py-ubjson==0.16.1
pyasn1==0.6.2
pyasn1-modules==0.4.2
pycodestyle==2.14.0
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pyflakes==3.4.0
pygments==2.19.2
pyjwt==2.11.0
pyopenssl==25.3.0
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-django==4.12.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.2
python-jose==3.5.0
pytokens==0.4.1
pyyaml==6.0.3
redis==7.2.1
requests==2.32.5
roman-numerals==4.1.0
rsa==4.9.1
service-identity==24.2.0
six==1.17.0
snowballstemmer==3.0.1
sphinx==9.1.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlalchemy==2.0.48
sqlparse==0.5.5
stack-data==0.6.3
starlette==0.52.1
traitlets==5.14.3
twisted==25.5.0
txaio==25.12.2
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
tzlocal==5.3.1
ujson==5.11.0
urllib3==2.6.3
uvicorn==0.41.0
vine==5.1.0
watchfiles==1.1.1
wcwidth==0.6.0
websockets==16.0
zope-interface==8.2
